[TNT-143] feat: App, AppState 최초 생성
