+++ feature/main/src/main/java/co/kr/tnt/main/ui/TnTApp.kt
@@ -1,0 +1,29 @@
+package co.kr.tnt.main.ui
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+@Suppress("UnusedParameter")
+fun TnTApp(
+    appState: TnTAppState,
+) {
+    Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+        Greeting(
+            name = "Trainer & Trainee",
+            modifier = Modifier.padding(innerPadding),
+        )
+    }
+}
+
+@Composable
+fun Greeting(name: String, modifier: Modifier = Modifier) {
+    Text(
+        text = "Hello $name!",
+        modifier = modifier,
+    )
+}
