+++ feature/main/src/main/java/co/kr/tnt/main/ui/TnTAppState.kt
@@ -1,0 +1,13 @@
+package co.kr.tnt.main.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.remember
+
+@Composable
+fun rememberTnTAppState(): TnTAppState {
+    return remember { TnTAppState() }
+}
+
+@Stable
+class TnTAppState
